Move transfer_tensors_to_device into utils/dist_utils
The helper is generic dataclass-tensor plumbing that distributed setup
code outside run_solver also needs. Lift it into utils/dist_utils so
those callers don't have to import from run_solver, and keep run_solver
using the same function via that import.
